Make the package field optional
in wasmer.toml, the package field will now be optional. This is to add support
for unnamed packages.

Unnamed packages are useful when you don't want to publish and maintain a
package just for an app. In that case, you can simply associate the package with
a webc hash, instead associating it with a package name/version etc.